+++ lib-impl/src/main/java/org/ergoplatform/appkit/impl/Eip29AttachmentBuilder.java
@@ -5,7 +5,9 @@ import static org.ergoplatform.appkit.impl.Eip4TokenBuilder.getSerializedErgoVal
 import org.ergoplatform.ErgoBox;
 import org.ergoplatform.appkit.Eip29Attachment;
 import org.ergoplatform.appkit.ErgoValue;
-import org.ergoplatform.appkit.GenericEip29Attachment;
+import org.ergoplatform.appkit.Eip29GenericAttachment;
+import org.ergoplatform.appkit.Eip29MultiAttachment;
+import org.ergoplatform.appkit.Eip29PlainTextAttachment;
 import org.ergoplatform.appkit.TransactionBox;
 import org.ergoplatform.explorer.client.model.AdditionalRegisters;
 
@@ -15,12 +17,18 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
 public class Eip29AttachmentBuilder {
+    /**
+     * @return register number for attachment to use according EIP-29
+     */
+    public static int getAttachmentRegisterIndex() {
+        return ErgoBox.maxRegisters() - 1;
+    }
 
     /**
      * @return EIP-29 compliant attachment built from serialized ergo value
      */
     public static Eip29Attachment buildFromHexEncodedErgoValue(String hexEncodedErgoValue) {
-        return GenericEip29Attachment.createFromErgoValue(ErgoValue.fromHex(hexEncodedErgoValue));
+        return Eip29GenericAttachment.createFromErgoValue(ErgoValue.fromHex(hexEncodedErgoValue));
     }
 
     /**
@@ -29,7 +37,7 @@ public class Eip29AttachmentBuilder {
     @Nullable
     public static Eip29Attachment buildFromAdditionalRegisters(@Nonnull AdditionalRegisters registers) {
         try {
-            return buildFromHexEncodedErgoValue(getSerializedErgoValueForRegister(registers, "R9"));
+            return buildFromHexEncodedErgoValue(getSerializedErgoValueForRegister(registers, "R" + getAttachmentRegisterIndex()));
         } catch (Throwable t) {
             return null;
         }
@@ -42,8 +50,8 @@ public class Eip29AttachmentBuilder {
     public static Eip29Attachment buildFromTransactionBox(@Nonnull TransactionBox transactionBox) {
         List<ErgoValue<?>> boxRegisters = transactionBox.getRegisters();
         try {
-            int register9Index = 9 - ErgoBox.startingNonMandatoryIndex();
-            return GenericEip29Attachment.createFromErgoValue(boxRegisters.get(register9Index));
+            int register9Index = getAttachmentRegisterIndex() - ErgoBox.startingNonMandatoryIndex();
+            return Eip29GenericAttachment.createFromErgoValue(boxRegisters.get(register9Index));
         } catch (Throwable t) {
             return null;
         }
@@ -52,14 +60,14 @@ public class Eip29AttachmentBuilder {
     /**
      * @return PlainTextAttachment for given text
      */
-    public static GenericEip29Attachment.PlainTextAttachment createPlainTextAttachment(String text) {
-        return GenericEip29Attachment.PlainTextAttachment.buildForText(text);
+    public static Eip29PlainTextAttachment createPlainTextAttachment(String text) {
+        return Eip29PlainTextAttachment.buildForText(text);
     }
 
     /**
      * @return MultiAttachment for given attachment list
      */
-    public static GenericEip29Attachment.MultiAttachment createMultiAttachment(List<Eip29Attachment> attachments) {
-        return GenericEip29Attachment.MultiAttachment.buildForList(attachments);
+    public static Eip29MultiAttachment createMultiAttachment(List<Eip29Attachment> attachments) {
+        return Eip29MultiAttachment.buildForList(attachments);
     }
 }
